add support for prevent_addition obj map attrib
the intention is that if you set prevent_addition, you  ca modify or
remove object maps but not create new objects.  not sure if this could
somehow effect non containing relations yet but the idea is to prevent a
new component from being created in a containing relation.  useful if
you use modelers for addition of objects but want to update them with
your python plugin.